// packages/oos-dsls-ts/src/generated/ast.ts
export interface AutoRefresh extends langium.AstNode {
    readonly $container: ViewDecl;
    readonly $type: 'AutoRefresh';
    domain: IdentName;
    event: IdentName;
}

export const AutoRefresh = {
    $type: 'AutoRefresh',
    domain: 'domain',
    event: 'event'
} as const;
// ...
export type BodyElement = Accordion | Button | Card | Divider | Grid | IconWidget | LinkWidget | RichText | Row | Section | Separator | Stack | Table | Tabs | Widget;

export const BodyElement = {
    $type: 'BodyElement'
} as const;
// ...
export interface DomainBinding extends langium.AstNode {
    readonly $container: ViewDecl;
    readonly $type: 'DomainBinding';
    alias?: IdentName;
    name: string;
}

export const DomainBinding = {
    $type: 'DomainBinding',
    alias: 'alias',
    name: 'name'
} as const;
// ...
export type ToolbarItem = ToolbarDelete | ToolbarExit | ToolbarNew | ToolbarRefresh | ToolbarSave;

export const ToolbarItem = {
    $type: 'ToolbarItem'
} as const;
// ...
export interface ViewDecl extends langium.AstNode {
    readonly $container: ViewModel;
    readonly $type: 'ViewDecl';
    autoRefresh?: AutoRefresh;
    body: Array<BodyElement>;
    default: boolean;
    domains: Array<DomainBinding>;
    name: string;
    title: string;
    toolbar: Array<ToolbarItem>;
}

export const ViewDecl = {
    $type: 'ViewDecl',
    autoRefresh: 'autoRefresh',
    body: 'body',
    default: 'default',
    domains: 'domains',
    name: 'name',
    title: 'title',
    toolbar: 'toolbar'
} as const;

export function isViewDecl(item: unknown): item is ViewDecl {
    return reflection.isInstance(item, ViewDecl.$type);
}

export interface ViewModel extends langium.AstNode {
    readonly $type: 'ViewModel';
    view: ViewDecl;
}

export const ViewModel = {
    $type: 'ViewModel',
    view: 'view'
} as const;